+++ packages/design-system/src/lib/components/form/form-group/form-element/FormSelect.tsx
@@ -8,10 +8,18 @@ export interface FormSelectProps
   isInvalid?: boolean
   isValid?: boolean
   disabled?: boolean
+  autoFocus?: boolean
 }
 
 export const FormSelect = React.forwardRef(function FormSelect(
-  { isInvalid, isValid, disabled, children, ...props }: PropsWithChildren<FormSelectProps>,
+  {
+    isInvalid,
+    isValid,
+    disabled,
+    autoFocus,
+    children,
+    ...props
+  }: PropsWithChildren<FormSelectProps>,
   ref
 ) {
   return (
@@ -19,6 +27,7 @@ export const FormSelect = React.forwardRef(function FormSelect(
       isInvalid={isInvalid}
       isValid={isValid}
       disabled={disabled}
+      autoFocus={autoFocus}
       ref={ref as React.ForwardedRef<HTMLSelectElement>}
       {...props}>
       {children}